Be careful with alloc-assign to ChimpValue attrs.
Wow. The copying collector will move values between heaps during a sweep,
so expressions like this:

  CHIMP_CODE(ref)->constants = chimp_array_new (NULL);

are not actually safe. This is because the address we want to assign to
(on the LHS) *will change* if the allocation triggers a sweep since the
value moves to the alternate heap. Here, we were saving the newly
allocated ChimpArray to the *previous* heap.

We use a temporary variable to work around the immediate issue here, but
I'm probably doing this stuff elsewhere, too.

Not sure what the C standard has to say about assignment involving
fields of values that may change location as a result of the
operations performed on the RHS... at best, it's completely unclear what
the 'correct' behaviour would be as a sane person reading this.